Change type of GPSOkBefore to float64
It appears that this is actually a float value. I encountered this error
when parsing real data:

json: cannot unmarshal number 1779296558.4 into Go struct field Aircraft.aircraft.gpsOkBefore of type int64

+++ aircraft.go
@@ -170,5 +170,5 @@ type Aircraft struct {
 
 	// GPS status (experimental).
 	// GPSOkBefore is the timestamp at which the GPS was last working well, if the aircraft has lost GPS or its GPS is heavily degraded. This is only included for 15 min after GPS is lost / degraded. This is experimental and subject to change.
-	GPSOkBefore *int64 `json:"gpsOkBefore"`
+	GPSOkBefore *float64 `json:"gpsOkBefore"`
 }
